Added UserCreateHikeTest and AdminCreateHikeTest (using login).
