--- MyHike/src/test/java/CreateHikeTest.java
@@ -1,73 +1,0 @@
-// Generated by Selenium IDE
-
-import database.Database;
-import facade.JPAFacade;
-import models.Hike;
-import org.junit.Test;
-import org.junit.Before;
-import org.junit.After;
-
-import static org.mockito.Mockito.*;
-
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.Dimension;
-import org.openqa.selenium.JavascriptExecutor;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.WebDriverWait;
-
-import java.time.Duration;
-import java.util.*;
-
-public class CreateHikeTest {
-    private WebDriver driver;
-    private Map<String, Object> vars;
-    JavascriptExecutor js;
-
-    @Before
-    public void setUp() {
-        driver = new ChromeDriver();
-        js = (JavascriptExecutor) driver;
-        vars = new HashMap<String, Object>();
-    }
-
-    @After
-    public void tearDown() {
-        driver.quit();
-    }
-
-    @Test
-    public void createHike() {
-        //Attempting to mock db
-        Database.facade = mock(JPAFacade.class);
-        doNothing().when(Database.facade).insert(any(Hike.class));
-
-        driver.get("http://localhost:8080/MyHike_war_exploded/discover.jsp");
-        driver.manage().window().setSize(new Dimension(1936, 1056));
-        WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(10)); // 10 seconds timeout
-        wait.until(ExpectedConditions.presenceOfElementLocated(By.linkText("Create Hike")));
-        driver.findElement(By.linkText("Create Hike")).click();
-        wait.until(ExpectedConditions.presenceOfElementLocated(By.id("map")));
-        driver.findElement(By.id("map")).click();
-        driver.findElement(By.id("map")).click();
-        driver.findElement(By.id("map")).click();
-        driver.findElement(By.id("name")).click();
-        driver.findElement(By.id("name")).sendKeys("Selenium Test");
-        driver.findElement(By.id("altitude")).click();
-        driver.findElement(By.id("altitude")).sendKeys("8");
-        driver.findElement(By.id("distance")).sendKeys("8");
-        driver.findElement(By.id("duration")).sendKeys("08:00");
-        driver.findElement(By.id("July")).click();
-        driver.findElement(By.id("August")).click();
-        driver.findElement(By.id("September")).click();
-        driver.findElement(By.cssSelector(".landscape-rating:nth-child(2) > .svg-inline--fa")).click();
-        driver.findElement(By.cssSelector(".strength-rating:nth-child(2) path")).click();
-        driver.findElement(By.cssSelector(".stamina-rating:nth-child(10) path")).click();
-        driver.findElement(By.cssSelector(".difficulty-rating:nth-child(10) > .svg-inline--fa")).click();
-        driver.findElement(By.id("description")).click();
-        driver.findElement(By.id("description")).sendKeys("test");
-        driver.findElement(By.id("fileToUpload")).sendKeys("C:\\Users\\Michael Reinprecht\\Desktop\\PXL_20231113_092920856.jpg");
-        driver.findElement(By.cssSelector(".btn")).click();
-    }
-}
